+++ Core/UserDefaultsPropertyWrapper.swift
@@ -170,7 +170,6 @@ public struct UserDefaultsWrapper<T> {
         // Debug keys
         case debugNewTabPageSectionsEnabledKey = "com.duckduckgo.ios.debug.newTabPageSectionsEnabled"
         case debugOnboardingHighlightsEnabledKey = "com.duckduckgo.ios.debug.onboardingHighlightsEnabled"
-        case debugNewTabPageSectionsBingSubKey = "com.duckduckgo.ios.debug.newTabPageSectionsBingSubKey"
 
         // Duck Player Pixel Experiment
         case duckPlayerPixelExperimentInstalled = "com.duckduckgo.ios.duckplayer.pixel.experiment.installed.v2"
